+++ src/util/KVStore.cpp
@@ -2,7 +2,7 @@
 
 template <typename T>
 KVStore<T>::KVStore(size_t shardCount) : shardCount(shardCount) {
-    for (int i = 0; i < shardCount; i++) {
+    for (size_t i = 0; i < shardCount; i++) {
         shards.emplace_back(std::make_unique<Shard>());
     }
 }
